requirements.txt
prometheus_flask_exporter
alembic==1.7.7
aniso8601==9.0.1
apispec==5.1.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
flask-apispec==0.11.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.34
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.1
zipp==3.8.0
